SrcLoc

Description

This module contains types that relate to the positions of things in source files, and allow tagging of those things with locations

SrcLoc

data RealSrcLoc

Real Source Location

Represents a single point within a file

data SrcLoc

Source Location

Constructors

RealSrcLoc !RealSrcLoc

UnhelpfulLoc FastString

Constructing SrcLoc

mkSrcLoc :: FastString -> Int -> Int -> SrcLoc

mkRealSrcLoc :: FastString -> Int -> Int -> RealSrcLoc

mkGeneralSrcLoc :: FastString -> SrcLoc

Creates a "bad" RealSrcLoc that has no detailed information about its location

noSrcLoc :: SrcLoc

Built-in "bad" RealSrcLoc values for particular locations

generatedSrcLoc :: SrcLoc

Built-in "bad" RealSrcLoc values for particular locations

interactiveSrcLoc :: SrcLoc

Built-in "bad" RealSrcLoc values for particular locations

advanceSrcLoc :: RealSrcLoc -> Char -> RealSrcLoc

Move the RealSrcLoc down by one line if the character is a newline, to the next 8-char tabstop if it is a tab, and across by one character in any other case

Unsafely deconstructing SrcLoc

srcLocFile :: RealSrcLoc -> FastString

Gives the filename of the RealSrcLoc

srcLocLine :: RealSrcLoc -> Int

Raises an error when used on a "bad" RealSrcLoc

srcLocCol :: RealSrcLoc -> Int

Raises an error when used on a "bad" RealSrcLoc

SrcSpan

data RealSrcSpan

A RealSrcSpan delimits a portion of a text file. It could be represented by a pair of (line,column) coordinates, but in fact we optimise slightly by using more compact representations for single-line and zero-length spans, both of which are quite common.

The end position is defined to be the column after the end of the span. That is, a span of (1,1)-(1,2) is one character long, and a span of (1,1)-(1,1) is zero characters long.

Real Source Span

data SrcSpan

Source Span

A SrcSpan identifies either a specific portion of a text file or a human-readable description of a location.

Constructors

RealSrcSpan !RealSrcSpan

UnhelpfulSpan !FastString

Constructing SrcSpan

mkGeneralSrcSpan :: FastString -> SrcSpan

Create a "bad" SrcSpan that has not location information

mkSrcSpan :: SrcLoc -> SrcLoc -> SrcSpan

Create a SrcSpan between two points in a file

mkRealSrcSpan :: RealSrcLoc -> RealSrcLoc -> RealSrcSpan

Create a SrcSpan between two points in a file

noSrcSpan :: SrcSpan

Built-in "bad" SrcSpans for common sources of location uncertainty

wiredInSrcSpan :: SrcSpan

Built-in "bad" SrcSpans for common sources of location uncertainty

interactiveSrcSpan :: SrcSpan

Built-in "bad" SrcSpans for common sources of location uncertainty

srcLocSpan :: SrcLoc -> SrcSpan

Create a SrcSpan corresponding to a single point

realSrcLocSpan :: RealSrcLoc -> RealSrcSpan

combineSrcSpans :: SrcSpan -> SrcSpan -> SrcSpan

Combines two SrcSpan into one that spans at least all the characters within both spans. Returns UnhelpfulSpan if the files differ.

srcSpanFirstCharacter :: SrcSpan -> SrcSpan

Convert a SrcSpan into one that represents only its first character

Deconstructing SrcSpan

srcSpanStart :: SrcSpan -> SrcLoc

Returns the location at the start of the SrcSpan or a "bad" SrcSpan if that is unavailable

srcSpanEnd :: SrcSpan -> SrcLoc

Returns the location at the end of the SrcSpan or a "bad" SrcSpan if that is unavailable

realSrcSpanStart :: RealSrcSpan -> RealSrcLoc

realSrcSpanEnd :: RealSrcSpan -> RealSrcLoc

srcSpanFileName_maybe :: SrcSpan -> Maybe FastString

Obtains the filename for a SrcSpan if it is "good"

pprUserRealSpan :: Bool -> RealSrcSpan -> SDoc

Unsafely deconstructing SrcSpan

srcSpanFile :: RealSrcSpan -> FastString

srcSpanStartLine :: RealSrcSpan -> Int

srcSpanEndLine :: RealSrcSpan -> Int

srcSpanStartCol :: RealSrcSpan -> Int

srcSpanEndCol :: RealSrcSpan -> Int

Predicates on SrcSpan

isGoodSrcSpan :: SrcSpan -> Bool

Test if a SrcSpan is "good", i.e. has precise location information

isOneLineSpan :: SrcSpan -> Bool

True if the span is known to straddle only one line. For "bad" SrcSpan, it returns False

containsSpan :: RealSrcSpan -> RealSrcSpan -> Bool

Tests whether the first span "contains" the other span, meaning that it covers at least as much source code. True where spans are equal.

Located

type Located = GenLocated SrcSpan

type RealLocated = GenLocated RealSrcSpan

data GenLocated l e

We attach SrcSpans to lots of things, so let's have a datatype for it.

Constructors

L l e

Constructing Located

noLoc :: HasSrcSpan a => SrcSpanLess a -> a

mkGeneralLocated :: HasSrcSpan e => String -> SrcSpanLess e -> e

Deconstructing Located

getLoc :: HasSrcSpan a => a -> SrcSpan

unLoc :: HasSrcSpan a => a -> SrcSpanLess a

unRealSrcSpan :: RealLocated a -> a

getRealSrcSpan :: RealLocated a -> RealSrcSpan

Combining and comparing Located values

eqLocated :: (HasSrcSpan a, Eq (SrcSpanLess a)) => a -> a -> Bool

Tests whether the two located things are equal

cmpLocated :: (HasSrcSpan a, Ord (SrcSpanLess a)) => a -> a -> Ordering

Tests the ordering of the two located things

combineLocs :: (HasSrcSpan a, HasSrcSpan b) => a -> b -> SrcSpan

addCLoc :: (HasSrcSpan a, HasSrcSpan b, HasSrcSpan c) => a -> b -> SrcSpanLess c -> c

Combine locations from two Located things and add them to a third thing

leftmost_smallest :: SrcSpan -> SrcSpan -> Ordering

Alternative strategies for ordering SrcSpans

leftmost_largest :: SrcSpan -> SrcSpan -> Ordering

Alternative strategies for ordering SrcSpans

rightmost :: SrcSpan -> SrcSpan -> Ordering

Alternative strategies for ordering SrcSpans

spans :: SrcSpan -> (Int, Int) -> Bool

Determines whether a span encloses a given line and column index

isSubspanOf

Arguments

:: SrcSpan
The span that may be enclosed by the other

-> SrcSpan
The span it may be enclosed by

-> Bool

Determines whether a span is enclosed by another one

sortLocated :: HasSrcSpan a => [a] -> [a]

HasSrcSpan

class HasSrcSpan a where

A typeclass to set/get SrcSpans

Methods

composeSrcSpan :: Located (SrcSpanLess a) -> a

Composes a SrcSpan decoration with an undecorated syntactic entity to form its decorated variant

decomposeSrcSpan :: a -> Located (SrcSpanLess a)

Decomposes a decorated syntactic entity into its SrcSpan decoration and its undecorated variant

type family SrcSpanLess a

Determines the type of undecorated syntactic entities For most syntactic entities E, where source location spans are introduced by a wrapper construtor of the same syntactic entity, we have `SrcSpanLess E = E`. However, some syntactic entities have a different type compared to a syntactic entity `e :: E` may have the type `Located E` when decorated by wrapping it with `L sp e` for a source span sp.

dL :: HasSrcSpan a => a -> Located (SrcSpanLess a)

An abbreviated form of decomposeSrcSpan, mainly to be used in ViewPatterns

cL :: HasSrcSpan a => SrcSpan -> SrcSpanLess a -> a

An abbreviated form of composeSrcSpan, mainly to replace the hardcoded L

pattern LL :: HasSrcSpan a => SrcSpan -> SrcSpanLess a -> a

A Pattern Synonym to Set/Get SrcSpans

onHasSrcSpan :: (HasSrcSpan a, HasSrcSpan b) => (SrcSpanLess a -> SrcSpanLess b) -> a -> b

Lifts a function of undecorated entities to one of decorated ones

liftL :: (HasSrcSpan a, HasSrcSpan b, Monad m) => (SrcSpanLess a -> m (SrcSpanLess b)) -> a -> m b